+++ agents-cli/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @inkeep/agents-cli
 
+## 0.1.8
+
+### Patch Changes
+
+- 17c116d: Readme fix
+  - @inkeep/agents-manage-ui@0.1.8
+  - @inkeep/agents-core@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- 712f1ad: Add url variables
+- Updated dependencies [a5756dc]
+- Updated dependencies [8aff3c6]
+- Updated dependencies [a0d8b97]
+  - @inkeep/agents-core@0.1.7
+  - @inkeep/agents-manage-ui@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes