+++ agents-manage-ui/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @inkeep/agents-manage-ui
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [fe6187f]
+  - @inkeep/agents-run-api@0.1.8
+  - @inkeep/agents-manage-api@0.1.8
+  - @inkeep/agents-core@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [a5756dc]
+- Updated dependencies [8aff3c6]
+- Updated dependencies [a0d8b97]
+- Updated dependencies [652895f]
+  - @inkeep/agents-core@0.1.7
+  - @inkeep/agents-run-api@0.1.7
+  - @inkeep/agents-manage-api@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes