+++ agents-run-api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inkeep/agents-run-api
 
+## 0.1.8
+
+### Patch Changes
+
+- fe6187f: Fix templates loading
+  - @inkeep/agents-core@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- a5756dc: Update model config resolution
+- 652895f: Remove conflicing port variables
+- Updated dependencies [a5756dc]
+- Updated dependencies [8aff3c6]
+- Updated dependencies [a0d8b97]
+  - @inkeep/agents-core@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes